collab: Attach GeoIP country code to RPC sessions
