+++ crates/collab/src/api/events.rs
@@ -1,4 +1,5 @@
 use super::ips_file::IpsFile;
+use crate::api::CloudflareIpCountryHeader;
 use crate::{api::slack, AppState, Error, Result};
 use anyhow::{anyhow, Context};
 use aws_sdk_s3::primitives::ByteStream;
@@ -52,33 +53,6 @@ impl Header for ZedChecksumHeader {
 
         let bytes = hex::decode(checksum).map_err(|_| axum::headers::Error::invalid())?;
         Ok(Self(bytes))
-    }
-
-    fn encode<E: Extend<axum::http::HeaderValue>>(&self, _values: &mut E) {
-        unimplemented!()
-    }
-}
-
-pub struct CloudflareIpCountryHeader(String);
-
-impl Header for CloudflareIpCountryHeader {
-    fn name() -> &'static HeaderName {
-        static CLOUDFLARE_IP_COUNTRY_HEADER: OnceLock<HeaderName> = OnceLock::new();
-        CLOUDFLARE_IP_COUNTRY_HEADER.get_or_init(|| HeaderName::from_static("cf-ipcountry"))
-    }
-
-    fn decode<'i, I>(values: &mut I) -> Result<Self, axum::headers::Error>
-    where
-        Self: Sized,
-        I: Iterator<Item = &'i axum::http::HeaderValue>,
-    {
-        let country_code = values
-            .next()
-            .ok_or_else(axum::headers::Error::invalid)?
-            .to_str()
-            .map_err(|_| axum::headers::Error::invalid())?;
-
-        Ok(Self(country_code.to_string()))
     }
 
     fn encode<E: Extend<axum::http::HeaderValue>>(&self, _values: &mut E) {
@@ -413,7 +387,7 @@ pub async fn post_events(
     let Some(last_event) = request_body.events.last() else {
         return Err(Error::Http(StatusCode::BAD_REQUEST, "no events".into()))?;
     };
-    let country_code = country_code_header.map(|h| h.0 .0);
+    let country_code = country_code_header.map(|h| h.to_string());
 
     let first_event_at = chrono::Utc::now()
         - chrono::Duration::milliseconds(last_event.milliseconds_since_first_event);
